worker/requirements.txt
mysql-connector-python
ollama